requirements.txt
# Web Framework
fastapi==0.108.0
uvicorn[standard]==0.25.0

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Audio Processing
audio-separator==0.41.1
librosa==0.10.1
soundfile==0.12.1
numpy==1.26.3
aiofiles==23.2.1
zipstream-ng==1.7.1
onnxruntime==1.23.2

# CORS & Middleware
python-multipart==0.0.6

# Testing
pytest==7.4.3
pytest-asyncio==0.23.3
httpx==0.26.0

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
